refactor, moving functions into attr and htmlfontelemend fixes #9639 #9638
